% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDAschd.R
\name{getGroupIQLRGenes}
\alias{getGroupIQLRGenes}
\title{Select groupIQLR genes as denominators for groupIQLR transformations}
\usage{
getGroupIQLRGenes(x, group)
}
\arguments{
\item{group}{The group information used for groupIQLR.}

\item{dat}{The raw count scRNA-seq matrix (gene(row) x cell(column)).}
}
\value{
The selected genes.
}
\description{
This function selects group-based IQLR(inter-quartile log-ratio) genes as denominators for groupIQLR transformations.
}
\examples{
selected_genes <- getGroupIQLRGenes(raw_dat,group=metadata$celltype);
}
