% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDAschd.R
\name{getGroupLVHAGenes}
\alias{getGroupLVHAGenes}
\title{Select groupLVHA genes as denominators for groupLVHA transformations}
\usage{
getGroupLVHAGenes(x, group)
}
\arguments{
\item{group}{The group information used for groupLVHA.}

\item{dat}{The raw count scRNA-seq matrix (gene(row) x cell(column)).}
}
\value{
The selected genes.
}
\description{
This function selects group-based LVHA(low-variance-high-abundance) genes as denominators for LVHA transformations.
}
\examples{
selected_genes <- getGroupLVHAGenes(raw_dat,group=metadata$celltype);
}
