% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDAschd.R
\name{runCoDAschd_LogNorm}
\alias{runCoDAschd_LogNorm}
\title{Run various CoDA LR transformations on scRNA-seq with prior log-normalization}
\usage{
runCoDAschd_LogNorm(
  dat,
  method = "CLR",
  gene = NULL,
  group = NULL,
  log_normalized = TRUE
)
}
\arguments{
\item{dat}{The raw or log-normalized count scRNA-seq matrix (gene(row) x cell(column)).}

\item{method}{The CoDA log-ratio method used. Options included: CLR(default); IQLR; LVHA; mdCLR; manual(should provide gene list as denominator); groupIQLR(should provide group information); groupLVHA(should provide group information).}

\item{gene}{The gene list used as denominator. Should be used with method="manual".}

\item{group}{The group information used for groupIQLR or groupLVHA. Should be used with method="groupIQLR" or "groupLVHA".}

\item{log_normalized}{Whether the input data has been log-normalized. If not, the raw data will be log2-normalized followed by LR transformations. Default is True.}
}
\value{
The output log-ratio matrix.
}
\description{
This function transforms the prior-log-normalized scRNA-seq matrix to different CoDA log-ratio-based matrix as an approximation.
}
\examples{
CLR_dat <- runCoDAschd_LogNorm(raw_dat,method="CLR",log_normalized=FALSE);

norm_dat <- t(CLOSE(t(raw_dat))*10000);
norm_dat <- log2(norm_dat+1);
IQLR_dat <- runCoDAschd_LogNorm(norm_dat,method="IQLR",log_normalized=TRUE);

norm_dat <- CreateSeuratObject(counts=raw_dat,min.cells=0,min.features=0);
norm_dat <- NormalizeData(norm_dat,verbose=FALSE); #natural log
norm_dat <- as.matrix(norm_dat@assays$RNA@data);
groupIQLR_dat <- runCoDAschd_LogNorm(norm_dat,method="groupIQLR",group=metadata$celltype,log_normalized=TRUE)

HKGLR_dat <- runCoDAschd_LogNorm(raw_dat,method="manual",gene=c("GAPDH","UBC"),log_normalized=FALSE);
}
